testsite/requirements.txt
# testsite for djaodjin-extended-templates
beautifulsoup4==4.11.1            # 4.11.1 compatibles with Py3.10
Django==3.2.25 ; python_version < "3.9"
Django==4.2.25 ; python_version >= "3.9" and python_version < "3.12"
Django==5.2.7  ; python_version >= "3.12"
djangorestframework==3.14.0 ; python_version < "3.9"
djangorestframework==3.15.2 ; python_version >= "3.9"
django-storages==1.14.4           # 1.14.4 not compatible with Django5.1
djaodjin-deployutils==0.14.1
Jinja2==3.1.3                     # 3.1.2 compatibles with Py3.10
premailer==3.10.0
requests==2.31.0
WeasyPrint==56.1

# testsite-only
coverage==6.4.4
django-extensions==3.2.3
gunicorn==23.0.0
whitenoise==6.2.0

# prerequesites for `django-storages`
boto3==1.33.13                    # last version compatible with Py3.7
